Come on this journey with me!
I'm starting to learn Emacs. If you're bored out of your mind on a
Saturday, then you can follow along with me as I struggle my way through
this process.

First a bit of background on myself: I've been using Vim for about 5
years now -- it's comfortable to me. And that's a problem. I figure it's
time I learned something new, but that doesn't necessarily mean I want
to throw out all the experienced I've gained editing files in Vim. There
are things I like about Vim, namely modal editing.

But modal editing is a concept, not an exclusive feature that Vim has on
lockdown... Emacs can be taught to understand modal editing fingers such
as mine, or so I'm told.

My agenda for this process will go something like this:

1. Learn just enough Emacs commands to open, edit, and save files.
   Install packages.

2. Install Evil and learn how it works.

3. Install Solarized color scheme.

4. Learn how to open files quickly with some sort of Fuzzy Finder
   equivalent (something I used ctrlp on Vim for). Further, learn how
   to open multiple files at once and navigate between them.

5. Learn how to do split panes ("windows" in Vim), both vertically and
   horizontally.

6. Learn how to navigate directory trees, similar to what I had with
   NERDTree. Learn how to create, move, delete files and folders.

7. Find an Ack or Grep equivalent in Emacs so I can search through a
   project.

8. Undo history is a must.

9. Clipboard history, maybe? I think I prefer my editor have its own
   clipboard with history, but I understand Emacs shares the system Mac
   clipboard, and maybe that's ok. We'll see...

10. Install surround plugin equivalent, i.e. editing/swapping parens,
    quotes, etc.

...and I'll probably think of other things as I go along.

Come with me on this journey of discovery and discomfort, and let's see
if how it goes!

-Tim