Emacs can see colors now
This is getting easier; this step only took me about 5 minutes.

As I was about to commit this, I noticed that Emacs didn't have colors
for the commit diff when I ran `git commit -v`, so I read through
the drama-filled EmacsWiki page here:

http://www.emacswiki.org/emacs/GitCommitMode

I'm still not sure I made the right choice, but I went with the
`git-commit-mode` package, and it seems to work well. One nice touch is
that now I also get on-the-fly spell-check of my commit messages.