Add project drawer for directory tree navigation
I feel kinda bad about this. I really should learn ECB or dired (or
both), but I have stuff to do and NERDTree was something of which I had
only learned just enough to navigate, create, rename, and delete files
and folders. I used it like maybe twice a day. So, I punted and
installed neowin, which appears to be a fairly close to NERDTree.

I won't say much about it -- just that the main keys I need are:

* <leader>p - toggle the neowin sidebar (my mapping)
* <leader>P - open neowin and highlight the current file (my mapping)
* C-c C-n   - create a new file or folder
* C-c C-r   - rename (move) a file or folder
* C-c C-d   - delete a file or folder