Allow cursor to jump to wrap
This one deserves its own commit. Thanks to https://github.com/jrichter
for finding this for me!

When I'm moving right and I get to the end of a line, I want my cursor
to start at the beginning of the next line (not beep at me like I did
something wrong). Likewise, moving to the beginning of of a line wraps
the cursor to the end of the previous line.

Yay!